+++ src/changelog.md
@@ -4,6 +4,9 @@ You can find the release notes below.
 ----------------------
 
 ## v1.1.3
+### Added
+- Added "Invoice date" field to custom invoices.
+
 ### Changes
 - The files tab is now consistently before the activity tab on all entity pages.
 
@@ -11,6 +14,7 @@ You can find the release notes below.
 - Fixed month in list of expired invoices being wrong (namely the day number of tomorrow instead of the actual month).
 - Fixed being able to mark paid/cancelled invoices as irrecoverable.
 - Fixed templates not being updated when the Docker image is updated.
+- Fixed input validation and NaN-dates of date input fields.
 
 ## v1.1.2 (31-08-2021)
 ### Bugfixes